Refactoring MapView, add DragGesture SheetBottom.
