+++ WeatherProjectTestOrion/WeatherProjectTestOrionApp.swift
@@ -9,12 +9,12 @@ import SwiftUI
 
 @main
 struct WeatherProjectTestOrionApp: App {
-    var body: some Scene {
-        WindowGroup {
-            //      MainView()
-            SearchCitiesView()
-                .environmentObject(WeatherViewModel(weatherService: WeatherService()))
-                .environmentObject(CitiesListViewModel(weatherService: WeatherService()))
-        }
+  var body: some Scene {
+    WindowGroup {
+//      MainView()
+      SearchCitiesOnMapView()
+        .environmentObject(WeatherViewModel(weatherService: WeatherService()))
+        .environmentObject(CitiesListViewModel(weatherService: WeatherService()))
     }
+  }
 }
